requirements.txt
numpy==1.23.5
scikit-learn==0.24.1 # must be version 0.24.1
pandas==1.5.2
matplotlib==3.6.3
xgboost==1.7.4
PyYAML==6.0
torchinfo==1.7.2
openpyxl
seaborn